fix rANS extraction crash and >64KB window corruption
Extraction of level 6-9 archives crashed (first seen on NetBSD/sdf.org,
reproducible everywhere), and files larger than the 64KB sliding window
silently corrupted at every level. Four causes:

- cli: master COMPRESS records hardcoded method 1 while master data was
  compressed at opt.level, so rANS masters were fed to the Huffman
  decoder. Records now carry method 10 at levels 6-9; levels 2-5 keep
  method 1 for original UC2 Pro compatibility.

- decompress: decompressor_rans stopped at remaining == 0 without
  consuming the end-of-block pair and its 12 extra bits, leaving the
  bit cursor desynchronized; the next block-present read landed inside
  the EOB extras and parsed a phantom block. The loop now decodes all
  nsyms symbols and guards output writes instead.

- decompress: a refill read returning a single byte into an empty
  buffer let head overtake tail in bits_feed; the unsigned difference
  wrapped and head walked off the 4KB buffer (the actual segfault).
  The refill now loops until a full byte pair is available, and a
  sticky error flag stops the decoder treating negative bits_get
  returns as data.

- compress/decompress: chunk loads wrote linearly past the circular
  window edge, and the rANS decoder flushed output in one linear write
  that cannot express ring wrap. Loads are now capped at the edge and
  the decoder flushes incrementally in ring order.

Also: BCJ E8/E9 byte assembly no longer shifts promoted ints into the
sign bit, and the libarchive plugin uses timegm on NetBSD/OpenBSD/
DragonFly so DOS timestamps are not offset by the local timezone.

New cli_bigfile regression test (>128KB round-trip at L5 and L6); it
fails against the previous binary and passes now. Verified: 22/22
ctest including the DOSBox-X round-trip against original uc2pro.exe,
ASan/UBSan clean, and the full matrix on NetBSD 10 (sdf.org).

+++ lib/src/uc2_preprocess.c
@@ -18,8 +18,8 @@ int uc2_bcj_apply(uint8_t *data, size_t len)
 	if (len < 5) return 0;
 	for (size_t i = 0; i + 4 < len; i++) {
 		if (data[i] == 0xE8 || data[i] == 0xE9) {
-			int32_t rel = (int32_t)(data[i+1] | (data[i+2] << 8) |
-			              (data[i+3] << 16) | (data[i+4] << 24));
+			int32_t rel = (int32_t)((uint32_t)data[i+1] | ((uint32_t)data[i+2] << 8) |
+			              ((uint32_t)data[i+3] << 16) | ((uint32_t)data[i+4] << 24));
 			int32_t abs_addr = rel + (int32_t)(i + 5);
 			data[i+1] = (uint8_t)(abs_addr);
 			data[i+2] = (uint8_t)(abs_addr >> 8);
@@ -36,8 +36,8 @@ int uc2_bcj_revert(uint8_t *data, size_t len)
 	if (len < 5) return 0;
 	for (size_t i = 0; i + 4 < len; i++) {
 		if (data[i] == 0xE8 || data[i] == 0xE9) {
-			int32_t abs_addr = (int32_t)(data[i+1] | (data[i+2] << 8) |
-			                   (data[i+3] << 16) | (data[i+4] << 24));
+			int32_t abs_addr = (int32_t)((uint32_t)data[i+1] | ((uint32_t)data[i+2] << 8) |
+			                   ((uint32_t)data[i+3] << 16) | ((uint32_t)data[i+4] << 24));
 			int32_t rel = abs_addr - (int32_t)(i + 5);
 			data[i+1] = (uint8_t)(rel);
 			data[i+2] = (uint8_t)(rel >> 8);
